Add basic scaffolding for distributing the package
It doesn't appear that the released mypy supports _stub packages right now, so for now we have to add to MYPYPATH manually.

Something else to consider is how to sync versions with numpy, but we can deal with that later, if we start our numbers low enough.

This adds a very small number of stubs just to verify that installation maybe works.

Fixes #2